Il tipo di price in Product adesso è double e non più int, aggiunta classe Utility

+++ src/com/company/store/purchase/Cart.java
@@ -1,6 +1,5 @@
 package com.company.store.purchase;
 
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -39,10 +38,10 @@ public final class Cart {
         }
     }
 
-    int getTotal() {
+    double getTotal() {
         //calculate total price
 
-        int total = 0;
+        double total = 0;
         for (CartEntry entry : contents.values()) {
             total += entry.getPrice();
         }
